server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=46.0.3",
    "fastapi>=0.127.0",
    "motor>=3.3.0",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "redis>=5.0.0",
    "hiredis>=2.2.0",
    "sqlalchemy>=2.0.45",
    "uvicorn[standard]>=0.40.0",
    "psycopg[binary]>=3.3.2",
    "pgvector>=0.3.0",
    "openai>=2.14.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.0.0,<5.0.0",
    "email-validator>=2.3.0",
    "apscheduler>=3.11.2",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]



[tool.pyright]
venvPath = "."
venv = ".venv"
pythonVersion = "3.12"
typeCheckingMode = "basic"

[project.scripts]
start = "python run.py"
